using size_t

+++ chapters/sorting_searching/bubble/code/c/bubble_sort.c
@@ -1,10 +1,11 @@
 #include <stdio.h>
+#include <stddef.h>
 
 void bubble_sort(int *array, size_t n) {
     int swapped = 0;
-    for (int i = 0; i < n - 1; ++i) {
+    for (size_t i = 0; i < n - 1; ++i) {
         swapped = 0;
-        for (int j = 0; j < n - i - 1; ++j) {
+        for (size_t j = 0; j < n - i - 1; ++j) {
             if (array[j] > array[j + 1]) {
                 int tmp = array[j];
                 array[j] = array[j + 1];
@@ -24,7 +25,7 @@ int main() {
     int array[10] = {1, 45, 756, 4569, 56, 3, 8, 5, -10, -4};
 
     printf("Unsorted array:\n");
-    for (int i = 0; i < 10; ++i) {
+    for (size_t i = 0; i < 10; ++i) {
         printf("%d ", array[i]);
     }
     printf("\n\n");
@@ -32,7 +33,7 @@ int main() {
     bubble_sort(array, 10);
 
     printf("Sorted array:\n");
-    for (int i = 0; i < 10; ++i) {
+    for (size_t i = 0; i < 10; ++i) {
         printf("%d ", array[i]);
     }
     printf("\n");
